+++ hippynn/__init__.py
@@ -24,3 +24,5 @@ from .experiment import setup_and_train
 from . import custom_kernels
 
 from . import pretraining
+
+from . import tools