+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = "2019, Los Alamos National Laboratory"
 author = "Nicholas Lubbers"
 
 # The full version, including alpha/beta/rc tags
-release = "0.0.1b4"
+release = "0.0.1"
 
 
 # -- General configuration ---------------------------------------------------